add 64-bit address support in uImage format (quick hack)

+++ tools/imagetool.h
@@ -45,8 +45,8 @@ struct image_tool_params {
 	int type;
 	int comp;
 	char *dtc;
-	unsigned int addr;
-	unsigned int ep;
+	unsigned long addr;
+	unsigned long ep;
 	char *imagename;
 	char *imagename2;
 	char *datafile;
